+++ src/main/java/com/github/galatynf/forglory/imixin/IAdrenalinMixin.java
@@ -1,6 +1,6 @@
 package com.github.galatynf.forglory.imixin;
 
 public interface IAdrenalinMixin {
-    double getAdrenalin();
-    void addAdrenalin(double amount);
+    float getAdrenalin();
+    void addAdrenalin(float amount);
 }
